+++ app/models/queryable_courses.rb
@@ -1,6 +1,6 @@
 class QueryableCourses
   def self.fetch(campus, term, rails_cache = Rails.cache)
-    MemoryCache.fetch(cache_key(campus, term)) do
+    FastCache.fetch(cache_key(campus, term)) do
       rails_cache.fetch(cache_key(campus, term)) do
         Course.for_campus_and_term(campus, term).collect do |course|
           QueryableCourse.new(course)
