+++ config/environments/development.rb
@@ -9,9 +9,8 @@ Rails.application.configure do
   # Do not eager load code on boot.
   config.eager_load = false
 
-  # Show full error reports and disable caching.
+  # Show full error reports
   config.consider_all_requests_local       = true
-  config.action_controller.perform_caching = false
 
   # Don't care if the mailer can't send.
   #config.action_mailer.raise_delivery_errors = false
@@ -38,6 +37,5 @@ Rails.application.configure do
 
   # Raises error for missing translations
   # config.action_view.raise_on_missing_translations = true
-  config.action_dispatch.rack_cache = false
-  config.cache_store = :null_store
+  config.caching.use = false
 end
